Export 패치
1. GitHub에서 리포지토리 fork전,
패치 내용에 대한 Export 파일 적용.
