+++ Assets/Jinho/Jinho_interface.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1df89db773f8b374d8467bdd0e027f73
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
