+++ Assets/Jinho/Weapon/DefaultBullet.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &7233024481384809248
+--- !u!1 &7213154079774707790
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,11 +8,12 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 7233024481384809252}
-  - component: {fileID: 7233024481384809255}
-  - component: {fileID: 7233024481384809254}
-  - component: {fileID: 7233024481384809249}
-  - component: {fileID: 4450151709307087562}
+  - component: {fileID: 7213154079774707893}
+  - component: {fileID: 7213154079774707786}
+  - component: {fileID: 7213154079774707787}
+  - component: {fileID: 7213154079774707784}
+  - component: {fileID: 7213154079774707785}
+  - component: {fileID: 8643283476328210379}
   m_Layer: 0
   m_Name: DefaultBullet
   m_TagString: Untagged
@@ -20,13 +21,13 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &7233024481384809252
+--- !u!4 &7213154079774707893
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7233024481384809248}
+  m_GameObject: {fileID: 7213154079774707790}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0.0066, y: 0.9981, z: 3.8239}
   m_LocalScale: {x: 0.03, y: 0.03, z: 0.03}
@@ -35,21 +36,21 @@ Transform:
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!33 &7233024481384809255
+--- !u!33 &7213154079774707786
 MeshFilter:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7233024481384809248}
+  m_GameObject: {fileID: 7213154079774707790}
   m_Mesh: {fileID: 10207, guid: 0000000000000000e000000000000000, type: 0}
---- !u!23 &7233024481384809254
+--- !u!23 &7213154079774707787
 MeshRenderer:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7233024481384809248}
+  m_GameObject: {fileID: 7213154079774707790}
   m_Enabled: 1
   m_CastShadows: 1
   m_ReceiveShadows: 1
@@ -85,26 +86,26 @@ MeshRenderer:
   m_SortingLayer: 0
   m_SortingOrder: 0
   m_AdditionalVertexStreams: {fileID: 0}
---- !u!135 &7233024481384809249
+--- !u!135 &7213154079774707784
 SphereCollider:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7233024481384809248}
+  m_GameObject: {fileID: 7213154079774707790}
   m_Material: {fileID: 0}
   m_IsTrigger: 1
   m_Enabled: 1
   serializedVersion: 2
   m_Radius: 0.5
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &4450151709307087562
+--- !u!114 &7213154079774707785
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 7233024481384809248}
+  m_GameObject: {fileID: 7213154079774707790}
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 355703f878f18e743aad8454cf0e577c, type: 3}
@@ -114,3 +115,16 @@ MonoBehaviour:
   damage: 0
   parentWeaponData: {fileID: 0}
   player: {fileID: 0}
+--- !u!114 &8643283476328210379
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7213154079774707790}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5dc6e7e291522f64fa0dd4fdabc504c7, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  poolingType: 0
