+++ Assets/Jinho/Weapon/Handgun_Prototype.prefab
@@ -201,7 +201,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   weaponData: {fileID: 11400000, guid: 442772fcfd3acaf4ba36776895ab3457, type: 2}
   firePos: {fileID: 7913764786492178657}
-  bullet: {fileID: 7233024481384809248, guid: 4eefa936c47806246969aca9484bf1d4, type: 3}
+  bullet: {fileID: 7213154079774707790, guid: 4eefa936c47806246969aca9484bf1d4, type: 3}
   maxBullet: 8
   bulletCount: 8
   maxTotalBullet: 64
