--- Assets/PolygonApocalypse/URP_ExtractMe.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: fbffd663618ba3f4e9a18c907f7dba12
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
